# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CueStats — Fremont Open API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: string;
  name: string;
  wins: number;
  losses: number;
  win_rate: number;
  /** @nullable */
  fargo?: number | null;
}

export interface Match {
  id: string;
  tournament_id: number;
  /** @nullable */
  tournament_name?: string | null;
  /** @nullable */
  round?: number | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  scores?: string | null;
  /** @nullable */
  winner_name?: string | null;
  /** @nullable */
  loser_name?: string | null;
  /** @nullable */
  completed_at?: string | null;
}

export interface Stats {
  total_tournaments: number;
  total_matches: number;
  total_players: number;
  /** @nullable */
  last_synced_at?: string | null;
  players: Player[];
  recent_matches: Match[];
}

export interface SyncStatus {
  status: string;
  /** @nullable */
  last_synced_at?: string | null;
  /** @nullable */
  tournaments_synced?: number | null;
  /** @nullable */
  matches_synced?: number | null;
  /** @nullable */
  players_rebuilt?: number | null;
}

export interface Tournament {
  id: number;
  name: string;
  /** @nullable */
  game?: string | null;
  state: string;
  /** @nullable */
  started_at?: string | null;
}

export interface TournamentDetail {
  tournament: Tournament;
  matches: Match[];
}

export interface H2HRecord {
  opponent: string;
  wins: number;
  losses: number;
}

export interface PlayerDetail {
  player: Player;
  matches: Match[];
  head_to_head: H2HRecord[];
}

export type PlayerExtrasStreaks = { [key: string]: unknown };

export type PlayerExtrasTitles = { [key: string]: unknown };

export type PlayerExtrasPerfVsFargo = { [key: string]: unknown };

export type PlayerExtrasWinsOverTimeItem = { [key: string]: unknown };

export interface PlayerExtras {
  streaks: PlayerExtrasStreaks;
  titles: PlayerExtrasTitles;
  perf_vs_fargo?: PlayerExtrasPerfVsFargo;
  wins_over_time: PlayerExtrasWinsOverTimeItem[];
  /** @nullable */
  fargo?: number | null;
}

export interface SearchResult {
  players: Player[];
  tournaments: Tournament[];
}

export type CompareResultH2h = {
  a_wins: number;
  b_wins: number;
  matches: Match[];
};

export type CompareResultCommonOpponentsItem = { [key: string]: unknown };

export interface CompareResult {
  a: Player;
  b: Player;
  h2h: CompareResultH2h;
  common_opponents: CompareResultCommonOpponentsItem[];
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminToken {
  token: string;
  email: string;
  role: string;
}

export interface AdminMe {
  email: string;
  role: string;
}

export interface SyncTriggerInput {
  force?: boolean;
  /** @nullable */
  tournament_id?: number | null;
}

export interface RenamePlayerInput {
  new_name: string;
}

export interface MergePlayersInput {
  canonical_name: string;
  alias_names: string[];
}

export interface FargoInput {
  /** @nullable */
  fargo?: number | null;
}

export interface MatchUpdateInput {
  /** @nullable */
  winner_name?: string | null;
  /** @nullable */
  loser_name?: string | null;
  /** @nullable */
  scores?: string | null;
}

export interface AdminActionResult {
  message: string;
  /** @nullable */
  matches_updated?: number | null;
}

export type AuditLogEntryPayload = { [key: string]: unknown };

export interface AuditLogEntry {
  id: number;
  action: string;
  payload?: AuditLogEntryPayload;
  created_at: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListPlayersParams = {
  q?: string;
};

export type GetLeaderboardParams = {
  limit?: number;
};

export type ListMatchesParams = {
  limit?: number;
};

export type SearchParams = {
  q: string;
  limit?: number;
};