# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CueStats — Fremont Open API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List players with optional search
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /players/{name}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player profile with matches and head-to-head
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        win_rate:
          type: number
        fargo:
          type: ["integer", "null"]
      required:
        - id
        - name
        - wins
        - losses
        - win_rate
    PlayerDetail:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/Player"
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
        head_to_head:
          type: array
          items:
            $ref: "#/components/schemas/H2HRecord"
      required:
        - player
        - matches
        - head_to_head
    H2HRecord:
      type: object
      properties:
        opponent:
          type: string
        wins:
          type: integer
        losses:
          type: integer
      required:
        - opponent
        - wins
        - losses
...
    Match:
      type: object
      properties:
        id:
          type: string
        tournament_id:
          type: integer
        tournament_name:
          type: ["string", "null"]
        round:
          type: ["integer", "null"]
        state:
          type: ["string", "null"]
        scores:
          type: ["string", "null"]
        winner_name:
          type: ["string", "null"]
        loser_name:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
      required:
        - id
        - tournament_id
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...