# lib/api-zod/src/generated/types/playerDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CueStats — Fremont Open API
 * OpenAPI spec version: 0.1.0
 */
import type { H2HRecord } from "./h2HRecord";
import type { Match } from "./match";
import type { Player } from "./player";

export interface PlayerDetail {
  player: Player;
  matches: Match[];
  head_to_head: H2HRecord[];
}